天穹飞控
增加飞行速度估计，暂时先写个基本功能，后续还需要上机测试，以及性能优化。
导航精度的提升，重点在于提高姿态估计的精度，并找到比较合理的误差模型，实现自适应融合

+++ SRC/SENSOR/barometer.h
@@ -5,7 +5,7 @@
 
 void BaroDataPreTreat(void);
 int32_t BaroGetAlt(void);
-int32_t BaroGetVelocity(void);
+float BaroGetVelocity(void);
 
 #endif
